Implement various fixed-point fixes.
darcs-hash:20071023044115-d1b56-708866692565e604a826276d34db3ad1e0385ae9.gz

+++ ale_accum.h
@@ -123,7 +123,7 @@ public:
 			if (M >= N)
 				i64_result = bits << (i64) ((int) M - (int) N);
 			else
-				i64_result = bits / (1 << (i64) ((int) N - (int) M));
+				i64_result = bits / ((i64) 1 << (i64) ((int) N - (int) M));
 
 			if (i64_result > ALE_FIXED_POSINF)
 				result.bits = ALE_FIXED_POSINF;
